requirements.txt
Django==2.2.27
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mysqlclient==1.3.14
Pillow==9.1.0
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
sqlparse==0.4.2